New implementation of password export
In the previous version the username was only available as part of the
password store identifier. Now it is also part of the content of the
store. The current version outputs the password as first line and the
username as second line. So it's now possible to use the "-c" option of
pass to selectively copy the first or second line to clipboard.

In addition the username is now only used as part of the password
identifier if more than one set of credentials exist for the same
address/url.

Keep in mind that exporting passwords will overwrite any existing
credentials that match the generated identifiers without prompting.