Add `--pass-compat=browserpass` to make --export-pass format compatible with browserpass
[browserpass](https://github.com/dannyvankooten/browserpass) expects a multiline format
with password first, and then a line with either `login:`, `user:` or `username:`.
Exporting passwords this way makes login form filling work if you have
browserpass installed.

To keep backwards compatibility the default, if `--pass-compat` is not
specified (or specified as `--pass-compat=default`), is to print just
the bare username on the second line.